+++ src/Queries/CacheModelQuery.php
@@ -6,7 +6,7 @@ use Illuminate\Database\Eloquent\Model;
 use Sfneal\Caching\Traits\Cacheable;
 use Sfneal\Helpers\Redis\RedisCache;
 
-abstract class AbstractQueryCacheAttribute extends AbstractQuery
+class CacheModelQuery extends AbstractQuery
 {
     /**
      * Inherit cache methods.
@@ -38,20 +38,35 @@ abstract class AbstractQueryCacheAttribute extends AbstractQuery
      * QueryCacheAttribute constructor.
      *
      * @param int $model_key
+     * @param string|null $model
+     * @param string|null $attribute
      */
-    public function __construct(int $model_key)
+    public function __construct(int $model_key, string $model = null, string $attribute = null)
     {
         $this->model_key = $model_key;
+        $this->model = $this->model ?? $model;
+        $this->attribute = $this->attribute ?? $attribute;
     }
 
     /**
      * Retrieve a Service's title.
      *
-     * @return string
+     * @return Model|string
      */
-    public function execute(): string
+    public function execute()
     {
-        return $this->model::query()->find($this->model_key)->getAttribute($this->attribute);
+        // Retrieve the model
+        $model = $this->model::query()->find($this->model_key);
+
+        // Return the entire model if no attribute is set
+        if (is_null($this->attribute)) {
+            return $model;
+        }
+
+        // Return a specific model
+        else {
+            return $model->getAttribute($this->attribute);
+        }
     }
 
     /**
@@ -62,8 +77,9 @@ abstract class AbstractQueryCacheAttribute extends AbstractQuery
     public function cacheKey(): string
     {
         $table = (new $this->model)->getTable();
+        $key = "{$table}:{$this->model_key}";
 
-        return "{$table}:{$this->model_key}:{$this->attribute}";
+        return $key.(is_null($this->attribute) ? '' : ":{$this->attribute}");
     }
 
     /**